requirements.txt
python-telegram-bot==20.7
pandas
python-dotenv
openpyxl
google-api-python-client
google-auth
